replace `xwallpaper` with `hsetroot`
Xwallpaper does not seem to properly render higher-resolution(than
that of my screen)? images. I don't know the cause of this. Upon reading
its manpage, there does not seem to be a solution for this problem. So
I'm just going to switch to something else that works. I've tested
setting images not properly rendered by xwallpaper with hsetroot and
they look way better - as far as I can see anyway.